+++ Netkan/Services/IHttpService.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 
 namespace CKAN.NetKAN.Services
 {
@@ -6,5 +7,7 @@ namespace CKAN.NetKAN.Services
     {
         string DownloadPackage(Uri url, string identifier, DateTime? updated);
         string DownloadText(Uri url);
+
+        IEnumerable<Uri> RequestedURLs { get; }
     }
 }